pyproject.toml
[project]
name = "agoffice"
version = "0.1.0"
description = ""
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.12"
classifiers = [
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
]
dependencies = [
    "fastapi[standard]",
    "uvicorn",
    "httpx",
    "python-socketio>=5.11",
    "sqlmodel",
    "kubernetes",
    "nanoid",
    "psycopg2-binary",
    "agpyutils @ git+https://github.com/iwaag/agpyutils.git@main",
    "redis>=5.0",
    "sse-starlette>=2.0",
]

[build-system]
requires = ["uv_build>=0.10.3,<0.11.0"]
build-backend = "uv_build"

[tool.uv]
package = false
